Winning and Blocking
Made progress on the winning and blocking algorithms
Need to go through the array, determine which spaces are free, if they are, test if the AI would win.
If there are no winning moves for the AI, they should look to block using the same process (see if the space is free, and if that space would win).
If there are no wins or blocks, play randomly.
Encountered an Index out of bounds error during the win-check phase. Will assess tomorrow.